# api/openapi.yml
openapi: 3.0.3
info:
  title: MTS Online Shop
  version: 0.1.0
  description: REST API интернет-магазина
servers:
  - url: http://localhost:8080/api
    description: Dev Server
...
paths:
...
  /products/{id}:
    get:
      summary: Получить информацию о товаре
      tags: [Products]
      operationId: getProductById
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Информация о товаре
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /cart:
    get:
      summary: Получить содержимое корзины
      tags: [Cart]
      operationId: getCart
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Корзина пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"
...
  /cart/clear:
    delete:
      summary: Очистить корзину
      tags: [Cart]
      operationId: clearCart
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Корзина очищена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /orders:
    get:
      summary: Получить заказы текущего пользователя
      tags: [Orders]
      operationId: getOrders
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Список заказов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
...
  /orders/{orderId}/cancel:
    post:
      summary: Отменить заказ с возвратом денег
      tags: [Orders]
      operationId: cancelOrder
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Заказ отменен, деньги возвращены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
                
  # Admin APIs
  /admin/users:
    get:
      summary: Получить список всех пользователей
      tags: [Admin Users]
      operationId: adminGetUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Список пользователей
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
        "401":
          description: Не авторизован
        "403":
          description: Недостаточно прав (только ADMIN)
  /admin/users/{userId}:
    get:
      summary: Получить информацию о пользователе
      tags: [Admin Users]
      operationId: adminGetUser
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Информация о пользователе
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Не авторизован
        "403":
          description: Недостаточно прав (только ADMIN)
        "404":
          description: Пользователь не найден
  /admin/users/{userId}/cart:
    get:
      summary: Получить корзину пользователя
      tags: [Admin Users]
      operationId: adminGetUserCart
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Корзина пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"
        "401":
          description: Не авторизован
        "403":
          description: Недостаточно прав (только ADMIN)
        "404":
          description: Пользователь не найден
  /admin/users/{userId}/cart/clear:
    delete:
      summary: Очистить корзину пользователя
      tags: [Admin Users]
      operationId: adminClearUserCart
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Корзина пользователя очищена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Не авторизован
        "403":
          description: Недостаточно прав (только ADMIN)
        "404":
          description: Пользователь не найден
  /admin/orders:
    get:
      summary: Получить все заказы всех пользователей
      tags: [Admin Orders]
      operationId: adminGetAllOrders
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Список всех заказов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
        "401":
          description: Не авторизован
        "403":
          description: Недостаточно прав (только ADMIN)
  /admin/orders/{orderId}/cancel:
    post:
      summary: Отменить любой заказ с возвратом
      tags: [Admin Orders]
      operationId: adminCancelOrder
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Заказ отменен, деньги возвращены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Не авторизован
        "403":
          description: Недостаточно прав (только ADMIN)
        "404":
          description: Заказ не найден
  /admin/products:
    get:
      summary: Получить все товары (админ)
      tags: [Admin Products]
      operationId: adminGetProducts
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Список товаров
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
        "401":
          description: Не авторизован
        "403":
          description: Недостаточно прав (только ADMIN)
    post:
      summary: Создать новый товар
      tags: [Admin Products]
      operationId: adminCreateProduct
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Товар создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Неверные данные товара
        "401":
          description: Не авторизован
        "403":
          description: Недостаточно прав (только ADMIN)
  /admin/products/{productId}:
    put:
      summary: Обновить товар
      tags: [Admin Products]
      operationId: adminUpdateProduct
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Товар обновлен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Неверные данные товара
        "401":
          description: Не авторизован
        "403":
          description: Недостаточно прав (только ADMIN)
        "404":
          description: Товар не найден
    delete:
      summary: Удалить товар
      tags: [Admin Products]
      operationId: adminDeleteProduct
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Товар удален
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Не авторизован
        "403":
          description: Недостаточно прав (только ADMIN)
        "404":
          description: Товар не найден
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: "Операция выполнена успешно"
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Пакет интернета 5Gb"
        price:
          type: number
          format: double
          example: 199.99
    ProductListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
...
    CartItem:
      type: object
      properties:
        itemId:
          type: integer
          format: int64
          example: 123
        product:
          $ref: "#/components/schemas/Product"
    CartResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        userId:
          type: integer
          format: int64
          example: 2
        totalPrice:
          type: number
          format: double
          example: 599.97
        status:
          type: string
          enum: [CREATED, PAID, DELIVERED, CANCELLED]
          example: "PAID"
        createdAt:
          type: string
          format: date-time
          example: "2024-03-15T10:30:00Z"
    OrderResponse:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/CartItem"
    OrderListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderResponse"
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        login:
          type: string
          example: "user1"
        email:
          type: string
          format: email
          example: "user@example.com"
        name:
          type: string
          example: "Иван Иванов"
        role:
          type: string
          enum: [USER, ADMIN]
          example: "USER"
        createdAt:
          type: string
          format: date-time
          example: "2024-03-15T10:30:00Z"
    UserResponse:
      $ref: "#/components/schemas/User"
    UserListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UserResponse"
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
    CreateProductRequest:
      type: object
      required:
        - name
        - price
      properties:
        name:
          type: string
          example: "Новый товар"
        price:
          type: number
          format: double
          example: 299.99
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
          example: "Обновленный товар"
        price:
          type: number
          format: double
          example: 399.99
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT